requirements.txt
absl-py==0.6.1
alabaster==0.7.8
altgraph==0.16.1
ansible==2.6.4
antiorm==1.2.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.5
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.4.0
backports.functools-lru-cache==1.5
bcrypt==3.1.5
blinker==1.4
Brlapi==0.6.6
cached-property==1.5.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
configparser==3.5.0
cryptography==2.3.1
cupshelpers==1.0
cycler==0.10.0
db==0.1.1
defer==1.0.6
distro-info==0.18
Django==2.0.5
dnspython==1.15.0
docker==3.5.0
docker-compose==1.22.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dominate==2.3.5
enum34==1.1.6
eventlet==0.24.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-MySQLdb==0.2.0
Flask-SocketIO==3.1.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
gast==0.2.0
greenlet==0.4.15
grpcio==1.16.0
Guake==3.0.5
gunicorn==19.9.0
h5py==2.8.0
httplib2==0.9.2
idna==2.8
imutils==0.5.1
inflection==0.3.1
influxdb==5.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
kazam==1.4.5
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
macholib==1.11
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.1
mccabe==0.6.1
monotonic==1.5
more-itertools==6.0.0
mpmath==1.0.0
mysqlclient==1.3.14
numpy==1.16.2
oauth==1.0.1
olefile==0.45.1
opencv-python==3.4.3.18
pandas==0.24.1
paramiko==2.4.2
pbr==3.1.1
pefile==2018.8.8
pexpect==4.2.1
Pillow==5.3.0
pluggy==0.9.0
protobuf==3.6.1
py==1.8.0
pyasn1==0.4.5
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.4
pygobject==3.26.1
pyinotify==0.9.6
PyInstaller==3.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyRFC3339==1.0
pytest==4.3.0
python-apt==1.6.2
python-dateutil==2.8.0
python-debian==0.1.32
python-engineio==3.0.0
python-socketio==2.1.0
pytz==2018.9
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.13
Quandl==3.4.6
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
rpy2==3.0.1
ruamel.yaml==0.15.71
scikit-learn==0.20.2
scipy==1.2.1
SecretStorage==2.3.1
simple-salesforce==0.74.2
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12.0
sklearn==0.0
SQLAlchemy==1.2.15
sympy==1.3
system-service==0.3
systemd-python==234
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
texttable==0.9.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
utils==0.9.0
virtualenv==15.2.0
visitor==0.1.3
wadllib==1.3.2
websocket-client==0.53.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xkit==0.0.0
xortool==0.96
youtube-dl==2018.3.14
zope.interface==4.3.2
